requirements_lightweight.txt
accelerate==0.34.2
albumentations==1.4.15
attr==0.3.2
ConfigParser==7.2.0
contextlib2==21.6.0
cryptography==45.0.2
dataset_utils==0.0.48
diffusers==0.31.0
dill==0.3.8
docutils==0.21.2
easydict==1.13
faiss==1.5.3
filelock==3.18.0
fire==0.7.0
HTMLParser==0.0.2
importlib_metadata==8.5.0
ipdb==0.13.13
ipython==8.12.3
ipywidgets==8.1.7
Jinja2==3.1.6
jnius==1.1.0
keyring==25.6.0
loralib==0.1.2
mock==5.2.0
munch==4.0.0
numpy==2.2.6
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
peft==0.13.2
Pillow==11.2.1
protobuf==6.31.0
pyOpenSSL==25.1.0
PyYAML==6.0.1
PyYAML==6.0.2
railroad==0.5.0
redis==6.1.0
safetensors==0.5.3
scikit_learn==1.6.1
scipy==1.15.3
Sphinx==8.2.3
tensorboardX==2.6.2.2
tensorboardX==2.6.2.2
thread==2.0.5
timm==1.0.15
torch==2.3.1
torchvision==0.18.1
torchviz==0.0.2
tornado==6.5.1
tqdm==4.66.4
transformers==4.46.2
trl==0.12.1
urllib3_secure_extra==0.1.0
xmlrpclib==1.0.1
